Cmdline export fixes and modpack format
