+++ Core/Exporters/PlainTextExporter.cs
@@ -6,7 +6,7 @@ namespace CKAN.Exporters
     public sealed class PlainTextExporter : IExporter
     {
 
-        public void Export(IRegistryQuerier registry, Stream stream)
+        public void Export(RegistryManager manager, IRegistryQuerier registry, Stream stream)
         {
             using (var writer = new StreamWriter(stream))
             {
